fix: Fixed room tests

+++ routes/web.php
@@ -24,7 +24,7 @@ use App\Http\Controllers\PasswordController;
 */
 //Only Guest
 Route::middleware(['only.guest'])->group(function () {
-    Route::get('/', [Controller::class, 'init']);
+    Route::get('/', [Controller::class, 'init'])->name('login');
     Route::post('/login', [LoginController::class, 'authenticate'])->name('user.login');
     Route::get('/register', [UserController::class, 'register_form'])->name('user.register');
     Route::post('/register', [UserController::class, 'register']);
@@ -33,30 +33,34 @@ Route::middleware(['only.guest'])->group(function () {
     })->name('user.forgot_password');
 });
 //Only Auth
-Route::middleware(['only.auth'])->group(function () {
+Route::middleware('auth')->group(function ()
+{
     Route::get('/main', [Controller::class, 'dashboard'])->name('dashboard');
     Route::get('/logout', [LoginController::class, 'logout']);
     Route::post('/user/set_profile/{user_id}', [UserSettingsController::class, 'set_user_profile']);
     Route::post('/user/set_settings', [UserSettingsController::class, 'save_user_settings']);
+
+    // Room
+    Route::get('/room', [RoomController::class, 'get_user_rooms']); //JSON Response
+    Route::get('/room/{room_id}', [RoomController::class, 'load_room'])->middleware('room.guard');
+    Route::post('/room', [RoomController::class, 'save_room'])->middleware('room.friends');
+    Route::put('/room/{room_id}', [RoomController::class, 'update_room_status']);
+    //Room settings
+    Route::delete('/room/{room_id}', [RoomController::class, 'delete_room'])->middleware('admin.room.guard');
+    Route::put('/room/{room_id}/update', [RoomController::class, 'update'])->middleware('admin.room.guard');
+    Route::post('/room/{room_id}/invite', [RoomController::class, 'invite'])->middleware('admin.room.guard');
+    //Room image
+    Route::post('/room/{room_id}/upload', [RoomController::class, 'upload_room_profile'])->middleware('admin.room.guard');
+    Route::get('/room/{room_id}/get_image', [RoomController::class, 'get_room_profile'])->middleware('admin.room.guard');
+    Route::put('/room/{room_id}/revert', [RoomController::class, 'revert_room_profile'])->middleware('admin.room.guard');
 });
 
 // Friendship
 Route::get('/friendship', [FriendshipController::class, 'get_user_friends']);
 Route::post('/friendship', [FriendshipController::class, 'save_friendship']);
 Route::put('/friendship/{friend_id}', [FriendshipController::class, 'update_friendship_status']);
-// Room
-Route::get('/room', [RoomController::class, 'get_user_rooms']); //JSON Response
-Route::get('/room/{room_id}', [RoomController::class, 'load_room'])->middleware('room.guard');
-Route::post('/room', [RoomController::class, 'save_room'])->middleware('room.friends');
-Route::put('/room/{room_id}', [RoomController::class, 'update_room_status']);
-//Room settings
-Route::delete('/room/{room_id}', [RoomController::class, 'delete_room']);
-Route::put('/room/{room_id}/update', [RoomController::class, 'update']);
-Route::post('/room/{room_id}/invite', [RoomController::class, 'invite']);
-//Room image
-Route::post('/room/{room_id}/upload', [RoomController::class, 'upload_room_profile']);
-Route::get('/room/{room_id}/get_image', [RoomController::class, 'get_room_profile']);
-Route::put('/room/{room_id}/revert', [RoomController::class, 'revert_room_profile']);
+
+
 // Messages
 Route::post('/chat/message/{room_id}', [MessagesController::class, 'send']);
 Route::post('/chat/file/{room_id}', [MessagesController::class, 'upload']);
